[REFACTOR]: GroupCreation BottomSheet를 개별 modifier 파일로 분리
GroupCreationView에 인라인된 모임명/목적/분위기 BottomSheet 정의를
각각 ViewModifier 파일로 추출
