+++ Projects/Presentation/HomeFeature/Sources/GroupCreation/GroupCreationView.swift
@@ -37,54 +37,9 @@ struct GroupCreationView: View {
         .background(.white)
         // 모임 목적(theme tag) 목록 fetch
         .task { await store.send(.onAppear) }
-        .bottomSheet(
-            isPresented: $store.isGroupNameSheetPresented,
-            header: .init(
-                title: "모임 만들기",
-                onClose: { store.send(.groupNameSheetDismissed) }
-            ),
-            primaryButton: .init(
-                title: "완료",
-                isEnabled: store.isGroupNameDraftValid,
-                action: { store.send(.groupNameConfirmed) }
-            )
-        ) {
-            TextInput(
-                title: "모임명",
-                placeholder: "ex: 삼총사 모임, 26기 대학동기",
-                maxCount: 30,
-                text: $store.groupNameDraft
-            )
-        }
-        .bottomSheet(
-            isPresented: $store.isPurposeSheetPresented,
-            header: .init(
-                title: "모임 목적",
-                onClose: { store.send(.purposeSheetDismissed) }
-            ),
-            primaryButton: .init(
-                title: "등록하기",
-                isEnabled: store.isPurposeDraftValid,
-                action: { store.send(.purposeConfirmed) }
-            )
-        ) {
-            PurposeSheetContent(store: store)
-        }
-        .bottomSheet(
-            isPresented: $store.isAtmosphereSheetPresented,
-            header: .init(
-                title: "장소 분위기",
-                description: "원하는 분위기를 골라보세요 (최대 3개)",
-                onClose: { store.send(.atmosphereSheetDismissed) }
-            ),
-            primaryButton: .init(
-                title: "등록하기",
-                isEnabled: store.isAtmosphereDraftValid,
-                action: { store.send(.atmosphereConfirmed) }
-            )
-        ) {
-            AtmosphereSheetContent(store: store)
-        }
+        .groupNameSheet(store: store)
+        .purposeSheet(store: store)
+        .atmosphereSheet(store: store)
     }
 }
 
@@ -142,128 +97,6 @@ private struct GroupCreationContent: View {
         store.selectedAtmospheres.isEmpty
             ? "편안한"
             : store.selectedAtmospheres.map(\.rawValue).joined(separator: ", ")
-    }
-}
-
-// MARK: - AtmosphereSheetContent
-
-private struct AtmosphereSheetContent: View {
-    let store: StoreOf<GroupCreationFeature>
-
-    var body: some View {
-        VStack(spacing: 0) {
-            ForEach(PlaceAtmosphere.allCases) { atmosphere in
-                AtmosphereRow(
-                    title: atmosphere.rawValue,
-                    isSelected: store.atmosphereDraft.contains(atmosphere),
-                    onTap: { store.send(.atmosphereToggled(atmosphere)) }
-                )
-            }
-        }
-        .frame(maxWidth: .infinity, alignment: .leading)
-    }
-}
-
-// MARK: - AtmosphereRow
-
-private struct AtmosphereRow: View {
-    let title: String
-    let isSelected: Bool
-    let onTap: () -> Void
-
-    var body: some View {
-        Button(action: onTap) {
-            HStack(spacing: Spacing.spacing200) {
-                Checkbox(
-                    variant: .circle,
-                    state: isSelected ? .enabled : .disabled,
-                    size: .medium
-                )
-
-                BangawoText(title, textStyle: .bodyMedium)
-                    .foregroundStyle(Colors.gray900)
-                    .frame(maxWidth: .infinity, alignment: .leading)
-            }
-            .padding(.vertical, Spacing.spacing300)
-            .contentShape(Rectangle())
-        }
-        .buttonStyle(.plain)
-    }
-}
-
-// MARK: - PurposeSheetContent
-
-/// 모임 목적(theme tag) 선택 BottomSheet 내용.
-/// 서버에서 fetch 한 `store.themeTags`(모임 목적 목록)를 행으로 렌더링한다.
-private struct PurposeSheetContent: View {
-    let store: StoreOf<GroupCreationFeature>
-
-    var body: some View {
-        VStack(spacing: 0) {
-            ForEach(store.themeTags) { themeTag in
-                PurposeRow(
-                    themeTag: themeTag,
-                    isSelected: store.purposeDraft == themeTag,
-                    onTap: { store.send(.purposeSelected(themeTag)) }
-                )
-            }
-        }
-        .frame(maxWidth: .infinity, alignment: .leading)
-    }
-}
-
-// MARK: - PurposeRow
-
-/// 모임 목적(theme tag) 단일 행. `themeTag.code` 로 로컬 아이콘을 매핑하며,
-/// 매핑되는 아이콘이 없는 목적은 아이콘 없이 displayName 만 표시한다.
-private struct PurposeRow: View {
-    let themeTag: ThemeTag
-    let isSelected: Bool
-    let onTap: () -> Void
-
-    var body: some View {
-        Button(action: onTap) {
-            HStack(spacing: Spacing.spacing300) {
-                if let icon = PurposeIcon.image(for: themeTag.code) {
-                    icon
-                        .resizable()
-                        .scaledToFill()
-                        .frame(width: Metric.iconLength, height: Metric.iconLength)
-                        .clipShape(Circle())
-                }
-
-                BangawoText(themeTag.displayName, textStyle: .bodyMedium)
-                    .foregroundStyle(Colors.gray900)
-                    .frame(maxWidth: .infinity, alignment: .leading)
-
-                if isSelected {
-                    Checkbox(variant: .ghost, state: .enabled, size: .medium)
-                }
-            }
-            .padding(.vertical, Spacing.spacing300)
-            .contentShape(Rectangle())
-        }
-        .buttonStyle(.plain)
-    }
-
-    private enum Metric {
-        static let iconLength: CGFloat = 32
-    }
-}
-
-// MARK: - PurposeIcon
-
-/// 모임 목적(theme tag) `code` → 로컬 아이콘.
-/// 서버 응답에는 아이콘 정보가 없으므로 클라이언트가 code 로 아이콘을 매칭한다.
-/// 목적 아이콘 에셋은 `ic_purpose_{keyword}` 포맷으로, 목적별로 trailing keyword 만 다르다.
-/// 서버 code(대문자 스네이크)를 소문자로 변환해 같은 포맷의 에셋 이름을 만들고,
-/// 그 이름의 에셋이 존재할 때만 매칭한다(목적별 매직스트링 분기 없음).
-/// 대응 에셋이 없는 code(예: `STUDY`)는 아이콘 없이 displayName 만 표시된다.
-private enum PurposeIcon {
-    private static let assetPrefix = "ic_purpose_"
-
-    static func image(for code: String) -> Image? {
-        Image.assetIfExists(named: assetPrefix + code.lowercased())
     }
 }
